# contracts/openapi/torgnexa-v1.yaml
openapi: 3.1.0
info:
  title: TORGNEXA API
  version: 0.16.0
servers:
  - url: /api/v1
paths:
...
  /settings/identity-providers:
    get:
      operationId: listSettingsIdentityProviders
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Bounded tenant-scoped current provider configurations, content: {application/json: {schema: {$ref: '#/components/schemas/IdentityProviderSettingsPage'}}}}
        '403': {description: Caller lacks settings.identity_providers.read}
  /settings/identity-providers/{provider_id}:
    put:
      operationId: saveSettingsIdentityProviderDraft
      security: [{bearerAuth: []}]
      parameters:
        - {name: provider_id, in: path, required: true, schema: {type: string, pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'}}
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 1, maxLength: 128}}
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/IdentityProviderDraft'}}}}
      responses:
        '200': {description: Immutable draft revision stored; active revision is unchanged, content: {application/json: {schema: {$ref: '#/components/schemas/IdentityProviderSettings'}}}}
        '409': {description: Optimistic version conflict}
        '422': {description: Issuer or callback is outside the deployment allowlist or is SSRF-unsafe}
...
  /compliance/evaluate:
    post:
      operationId: evaluateProductCompliance
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExternalComplianceEvaluationRequestV1'}
      responses:
        '200':
          description: Explainable product-compliance decision for an authenticated tenant
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExternalComplianceEvaluationResultV1'}
        '400': {description: Invalid compliance evaluation request}
        '401': {description: Authenticated tenant scope missing}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    IdentityProviderSettings:
      type: object
      additionalProperties: false
      required: [provider_id, protocol, display_name, issuer_url, client_id, callback_url, secret_configured, revision, version, enabled, validation_status, validation_reason, updated_at]
      properties:
        provider_id: {type: string, pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'}
        protocol: {type: string, const: oidc}
        display_name: {type: string, minLength: 1, maxLength: 160}
        issuer_url: {type: string, format: uri, maxLength: 2048}
        client_id: {type: string, minLength: 1, maxLength: 256}
        callback_url: {type: string, format: uri, maxLength: 2048}
        secret_configured: {type: boolean}
        revision: {type: integer, minimum: 1}
        version: {type: integer, minimum: 1}
        active_revision: {type: integer, minimum: 1}
        enabled: {type: boolean}
        validation_status: {type: string, enum: [not_validated, valid, invalid]}
        validation_reason: {type: string, pattern: '^[a-z][a-z0-9._-]{0,63}$'}
        validated_at: {type: string, format: date-time, pattern: 'Z$'}
        updated_at: {type: string, format: date-time, pattern: 'Z$'}
    IdentityProviderSettingsPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/IdentityProviderSettings'}}
    IdentityProviderDraft:
      type: object
      additionalProperties: false
      required: [protocol, display_name, issuer_url, client_id, callback_url, expected_version]
      properties:
        protocol: {type: string, const: oidc}
        display_name: {type: string, minLength: 1, maxLength: 160}
        issuer_url: {type: string, format: uri, maxLength: 2048}
        client_id: {type: string, minLength: 1, maxLength: 256}
        callback_url: {type: string, format: uri, maxLength: 2048}
        client_secret: {type: string, format: password, maxLength: 65536, writeOnly: true}
        expected_version: {type: integer, minimum: 0}
...
    ExternalComplianceEvaluationRequestV1:
      additionalProperties: false
      title: Compliance Evaluation Request v1
      type: object
      required:
      - operation
      - jurisdiction
      - product_id
      - at
      properties:
        operation:
          enum:
          - publication
          - sale
          - advertising
          - shipping
        jurisdiction:
          type: string
          pattern: ^[A-Z]{2}$
        product_id:
          type: string
          pattern: ^(?:[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}|[0-7][0-9A-HJKMNP-TV-Z]{25})$
        offer_id:
          type: string
          pattern: ^(?:[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}|[0-7][0-9A-HJKMNP-TV-Z]{25})$
        category_id:
          type: string
          pattern: ^(?:[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}|[0-7][0-9A-HJKMNP-TV-Z]{25})$
        gtin:
          type: string
          pattern: ^(?:[0-9]{8}|[0-9]{12}|[0-9]{13}|[0-9]{14})$
        sku:
          type: string
          maxLength: 128
        seller_party_id:
          type: string
          pattern: ^(?:[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}|[0-7][0-9A-HJKMNP-TV-Z]{25})$
        seller_role:
          type: string
          maxLength: 128
        connector_id:
          type: string
          maxLength: 128
        connector_family:
          type: string
          maxLength: 128
        at:
          type: string
          format: date-time
          pattern: ^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:\.[0-9]+)?Z$
    ExternalComplianceEvaluationResultV1:
      additionalProperties: false
      title: Compliance Evaluation Result v1
      type: object
      required:
      - outcome
      - reasons
      - evaluated_at
      - fingerprint_sha256
      properties:
        outcome:
          enum:
          - allow
          - warn
          - approval_required
          - block
        reasons:
          type: array
          maxItems: 512
          items:
            type: object
            additionalProperties: false
            required:
            - code
            - policy_id
            - policy_version
            - document_type
            properties:
              code:
                enum:
                - missing_evidence
                - expired_evidence
                - unverified_evidence
                - invalid_evidence
              policy_id:
                type: string
                pattern: ^(?:[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}|[0-7][0-9A-HJKMNP-TV-Z]{25})$
              policy_version:
                type: integer
                minimum: 1
              document_type:
                enum:
                - declaration
                - certificate
                - eac_evidence
                - state_registration
                - veterinary
                - sanitary
                - refusal_letter
                - information_letter
                - other
              evidence_ids:
                type: array
                maxItems: 256
                items:
                  type: string
                  pattern: ^(?:[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}|[0-7][0-9A-HJKMNP-TV-Z]{25})$
        evaluated_at:
          type: string
          format: date-time
          pattern: ^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:\.[0-9]+)?Z$
        fingerprint_sha256:
          type: string
          pattern: ^[0-9a-f]{64}$
...